os_k8s_agent_system/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
faiss-cpu>=1.7.4
pypdf2>=3.0.1
pymupdf>=1.23.8
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
tabulate>=0.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
